Move CSS to external file and reset timer when a new slide is shown
It is good practice to move the CSS into an external file, and
honestly, I should have done that before. Better late than never.

Resetting the timer whenever a new slide is shown fixes a bug where
the carousel auto-advances while the user is interacting with it.
If the user is interacting with the carousel, the carousel should
not attempt to advance. It should only advance when the user does
not interact.