Allow users to control slideDuration via views.py and reset the timer when slides change
Allowing users to control slideDuration via views.py makes it easy to change the duration
of each slide without having to affect the actual JS logic.

Resetting the timer whenever the slide changes fixes a bug that can occur
whenever the slide auto-advances while the user interacts with the carousel.