Remove misleading buttons on gitcoin UI
We are currently generating some buttons on the gitcoin UI
by first querying information about the bounty and then deciding
what buttons to add on the UI based on that info.

However, we are only generating these buttons based on the
status of the bounty on the blockchain, and we do not take into account
whether the user is supposed to see those buttons. (For example, it makes
little sense for a person who just created a bounty to then claim their
own bounty.)

To prevent users from seeing buttons that they are not meant to see,
we will use the 'bounty_owner_address' to decide whether somebody
is a "Bounty Owner". If a user is currently logged into Metamask
and have the same address as the 'bounty_owner_address',
then they are likely the Bounty Owner. We then use this knowledge (in addition
to the current status of the bounty) to decide what buttons we should generate.

Bounty owners cannot claim their own issues.

Non-"bounty owners" cannot accept/reject issues or clawback funds.