Display random quote when a web3 transaction takes place
The hope is that this random quote will entertain users
while they are waiting for the transaction to be processed.

A random quote is selected and displayed to the user every 30
seconds until the transaction is fully processed. Future programmers
are free to add in more quotes or change the duration between each
quote.

Ref: https://github.com/gitcoinco/web/issues/32