Generate high-entropy secrets and codes
Gitcoin Core currently generates hashes from low-entropy, predictable
data. While they may appear random to outside observers, looking at the
source code quickly shows these as being predictable.

The construct used in both places appears to be a RIPEMD-160 hash
salted with the RIPEMD-160 hash for empty string and then the Django
timezone. An adversary with knowledge of the rough time in which a
secret was generated may use this knowledge to predict the hash used for
the whitepaper access code or the `priv` property.

This pull request removes the `hashlib` from use and generates
hexadecimal secrets from the OS `urandom` facility. These strings are
then truncated to match the same length as previously to avoid any
compatibility issues.

While the construct with `codecs` may appear strange, it will generate
the proper string in both Python 2.x and 3.x.

Testing: No joke, I just tested the construct, linted this and
called it good. Please pull this and try it locally before merge. I
don't have docker / docker-compose working on this workstation yet.