Fix broken running locally link
'Running locally' link in the Table of contents was broken as we didn't have a corresponding section with the href `running-locally`

This PR adds in a new heading, thereby creating a ref for the 'Running locally' link.

Also changed the level of the sub-headings and the text (With Docker, Without docker). Removed 'Running locally' before those headings as it seemed a bit redundant.