Use complete url query param value to resolve bounty's detail
We use a 'url' query param to resolve the bounty details of non-github bounties rather than the
ghuser, ghrepo, and ghissue id. As this 'url' query param value may contain query params relevant
to the non-github bounty, we must ensure the request_url variable is set to the complete 'url'
query param value.

Fixes: https://github.com/gitcoinco/web/issues/1822