Improve database syncing logic
Break up the syncing to handle a regular "sync", vs. a "transfer".  The "transfer" needs to access  the kudos_transfer model to figure out what the `send_from_address` is.

This is due to the Kudos Indirect Send method - We use a proxy account to store the funds and then do a raw transaction.  So `tx['from']` cannot be trusted.