Grant contribution activity (#3316)
When a user subscribes/contributes to a grant, or when she unsubscribes,
the event is recorded as a new Activity object. In order to accomplish
this, a new foreign key was added to the Activity object, which references
the grant subscriptions object. I experienced a few crashes after
implementing this change, although I'm not sure they're related.