Proposed fix for Kudos marketplace search bar
When you add whitespace in search bar value, for example: "Limited edition", the query parameters will be "Limited+edition", so when results come back in the view, the search bar get its value from query parameters (queryParam.q), so it will also show up in the search bar as Limited+edition. The simplest solution is to replace the plus signs with whitespace, i think it does look better.