update tab section
add a third item to the tab tuple to act as title in views.
previous version (title, count)
new version (id, title, count)

done so I do not have rework it too much. This way, the title can be variable
and can be renamed fairly simply while the the id remains constant.